pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dbt-job-maestro"
version = "0.1.0"
description = "Automatically generate dbt selectors from manifest.json by analyzing dependencies, paths, and tags"
readme = "README.md"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
license = { text = "MIT" }
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Database",
]
keywords = ["dbt", "selectors", "manifest", "data", "analytics", "automation"]
dependencies = [
    "pyyaml>=6.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]

[project.urls]
Homepage = "https://github.com/yourusername/dbt-job-maestro"
Repository = "https://github.com/yourusername/dbt-job-maestro"
Issues = "https://github.com/yourusername/dbt-job-maestro/issues"

[project.scripts]
dbt-job-maestro = "dbt_job_maestro.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["dbt_job_maestro*"]
exclude = ["tests*", "examples*"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
